+++ lib/screens/timer/src/timer_ring/src/clock_label.dart
@@ -1,0 +1,29 @@
+import 'package:auto_size_text/auto_size_text.dart';
+import 'package:flutter/material.dart';
+import 'package:preptime/models/debate_event.dart';
+import 'package:preptime/models/event.dart';
+import 'package:preptime/models/speech.dart';
+import 'package:provider/provider.dart';
+import 'package:preptime/utilities/duration_format/duration_format.dart';
+
+/// The amount of time left on the clock.
+class ClockLabel extends StatelessWidget {
+  ClockLabel({Key key}) : super(key: key);
+  @override
+  Widget build(BuildContext context) {
+    Speech speech = context.watch<Speech>();
+    Event event = context.watch<Event>();
+    bool isDisabled = (event is DebateEvent) && event.isAnyRunning;
+    return Container(
+      alignment: FractionalOffset.center,
+      padding: EdgeInsets.symmetric(horizontal: 20),
+      child: AutoSizeText(
+        speech.timeRemaining.toStringAsClock(),
+        maxLines: 1,
+        style: isDisabled
+            ? Theme.of(context).textTheme.headline2
+            : Theme.of(context).textTheme.headline1,
+      ),
+    );
+  }
+}