+++ lib/utilities/modals/modals.dart
@@ -1,0 +1,9 @@
+library modals;
+
+import 'dart:io';
+
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+
+part './src/clear_timer.dart';